+++ frontend/src/lib/common.ts
@@ -8,7 +8,7 @@ export type SupportedLanguage = Script.language
 
 export interface SchemaProperty {
 	type: string | undefined
-	description: string
+	description?: string
 	pattern?: string
 	default?: any
 	enum?: string[]