+++ frontend/src/lib/components/apps/editor/settingsPanel/ArrayStaticInputEditor.svelte
@@ -121,7 +121,7 @@
 	}
 </script>
 
-<div class="flex gap-2 flex-col mt-2">
+<div class="flex gap-2 flex-col mt-2 w-full">
 	{#key redraw}
 		{#if Array.isArray(items) && componentInput.value}
 			<div class="text-xs text-gray-600 font-semibold">{pluralize(items.length, 'item')}</div>
@@ -140,7 +140,9 @@
 						<!-- svelte-ignore a11y-no-noninteractive-tabindex -->
 
 						<div class="flex flex-row gap-2 items-center relative my-1">
-							<SubTypeEditor {subFieldType} bind:componentInput bind:value={item.value} />
+							<div class="grow overflow-x-auto">
+								<SubTypeEditor {subFieldType} bind:componentInput bind:value={item.value} />
+							</div>
 							<div class="flex justify-between flex-col items-center">
 								<div
 									tabindex={dragDisabled ? 0 : -1}