+++ frontend/src/lib/components/apps/editor/settingsPanel/script/utils.ts
@@ -1,4 +1,4 @@
-import type { Runnable } from '$lib/components/apps/inputType'
+import type { EvalAppInput, Runnable } from '$lib/components/apps/inputType'
 import type { AppComponent } from '../../component'
 import type {
 	StaticAppInput,
@@ -9,7 +9,7 @@ import type {
 
 export function getDependencies(
 	fields:
-		| Record<string, StaticAppInput | ConnectedAppInput | RowAppInput | UserAppInput>
+		| Record<string, StaticAppInput | ConnectedAppInput | RowAppInput | UserAppInput | EvalAppInput>
 		| undefined
 ): string[] {
 	let dependencies: string[] = []
@@ -20,6 +20,10 @@ export function getDependencies(
 		if (field.type === 'connected' && dependencies && field.connection) {
 			dependencies.push(`${field.connection.componentId} - ${field.connection.path}`)
 		}
+
+		if (field.type === 'eval' && !dependencies.includes('Eval')) {
+			dependencies.push(`Eval`)
+		}
 	})
 	return dependencies
 }
@@ -29,7 +33,9 @@ export function isFrontend(runnable: Runnable): boolean {
 }
 
 export function isTriggerable(componentType: string): boolean {
-	return ['buttoncomponent', 'formbuttoncomponent', 'formcomponent'].includes(componentType)
+	return ['buttoncomponent', 'formbuttoncomponent', 'formcomponent', 'steppercomponent'].includes(
+		componentType
+	)
 }
 
 export function isTriggerOnAppLoad(appComponent: AppComponent): boolean {